Fix queries if node class label is added
Prometheus keeps series "alive" for 5 minutes after they get dropped from a scrape target. This leads to a slight overlap in the `kube_node_labels` metric if a node is relabeled.

This commit prioritizes series having a node class labels over series without.

This still does not work if a node class label changes from non-empty to non-empty but that should never happen. I don't think there is a generic way to handle this case.